Fixing goofy type mistake
The time clock was limited to a 16 bit uint.  This meant that the
clock would reset at 2^16 milliseconds--which experienced players hit
around level 16 consistently.

+++ src/model/PlayModel.cpp
@@ -339,7 +339,7 @@ void PlayModel::setState(State state) {
     this->state = state;
 }
 
-uint16_t PlayModel::getTimeClock() {
+uint32_t PlayModel::getTimeClock() {
     return timeClock;
 }
